fix Item argument type

+++ src/main/java/org/vaibhav/cart/Item.java
@@ -4,7 +4,7 @@ public class Item {
     String name;
     double price;
 
-    public Item(String name, int price){
+    public Item(String name, double price){
         this.name = name;
         if(price > 0) this.price = price;
         else throw new IllegalArgumentException();
